requirements.txt
aiogram==2.13
aiohttp==3.7.4.post0
aiojobs==0.3.0
aiomysql==0.0.21
aioredis==1.3.1
async-class==0.3.0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
callback==0.0.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
dateparser==1.0.0
decorator==5.0.9
environs==9.3.2
hiredis==2.0.0
idna==2.10
loguru==0.5.3
marshmallow==3.12.2
multidict==5.1.0
mysql-connector-python==8.0.25
protobuf==3.17.3
pyaes==1.6.1
pyasn1==0.4.8
PyMySQL==0.9.3
python-binance==1.0.12
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
self==2020.12.3
six==1.16.0
Telethon==1.22.0
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
win32-setctime==1.0.3
yarl==1.6.3

Pyrogram~=1.2.9